+++ internal/check/metrics_test.go
@@ -70,7 +70,6 @@ func TestUpdateCheckBundleMetrics(t *testing.T) {
 
 	mc := minimock.NewController(t)
 	client := genMockClient(mc)
-
 	c := Check{bundle: &apiclient.CheckBundle{CID: ""}, client: client}
 
 	t.Log("nil metrics")
@@ -97,7 +96,7 @@ func TestUpdateCheckBundleMetrics(t *testing.T) {
 	{
 		c.bundle.CID = "/check_bundle/000"
 		metrics := map[string]apiclient.CheckBundleMetric{
-			"foo": {Name: "foo", Type: "n", Status: "active"},
+			"foo": apiclient.CheckBundleMetric{Name: "foo", Type: "n", Status: "active"},
 		}
 		err := c.updateCheckBundleMetrics(&metrics)
 		if err == nil {
@@ -112,7 +111,7 @@ func TestUpdateCheckBundleMetrics(t *testing.T) {
 	{
 		c.bundle.CID = "/check_bundle/0002"
 		metrics := map[string]apiclient.CheckBundleMetric{
-			"foo": {Name: "foo", Type: "n", Status: "active"},
+			"foo": apiclient.CheckBundleMetric{Name: "foo", Type: "n", Status: "active"},
 		}
 		err := c.updateCheckBundleMetrics(&metrics)
 		if err == nil {
@@ -127,7 +126,7 @@ func TestUpdateCheckBundleMetrics(t *testing.T) {
 	{
 		c.bundle.CID = "/check_bundle/1234"
 		metrics := map[string]apiclient.CheckBundleMetric{
-			"foo": {Name: "foo", Type: "n", Status: "active"},
+			"foo": apiclient.CheckBundleMetric{Name: "foo", Type: "n", Status: "active"},
 		}
 		err := c.updateCheckBundleMetrics(&metrics)
 		if err != nil {
